requirements.txt
asgiref==3.8.1
crispy-bootstrap5==2024.2
Django==5.1.1
django-crispy-forms==2.3
django-filter==24.3
djangorestframework==3.15.2
Markdown==3.7
psycopg2-binary==2.9.9
python-dotenv==1.0.1
sqlparse==0.5.1
typing_extensions==4.12.2
